Extract out the responsibility of Formatting text for display from TextUi in to a separate class named Formatter
